refactor: Remove old trait and dependencies
