+++ crates/hyperswitch_interfaces/src/lib.rs
@@ -1,7 +1,5 @@
 //! Hyperswitch interface
 
-#![warn(missing_docs, missing_debug_implementations)]
-
 pub mod api;
 
 pub mod configs;
